Remove h2m from Yari
