+++ package.json
@@ -23,7 +23,8 @@
     "dev": "yarn build:prepare && nf -j Procfile.dev start",
     "eslint": "eslint .",
     "filecheck": "cd filecheck && ts-node cli.ts",
-    "md": "ts-node markdown/cli.ts",
+    "m2h": "ts-node markdown/m2h/cli.ts",
+    "md": "echo 'The HTML to Markdown conversion command has been moved to https://github.com/mdn/markdown!' && exit 1",
     "prepack": "yarn build:dist",
     "prepare": "husky install",
     "prettier-check": "prettier --check .",
