+++ scripts/testing.sh
@@ -76,11 +76,9 @@ sleep 1
     echo "----------"
 }
 
-echo "--------------------------"
-echo "Basic m2h/h2m tool testing"
-echo "--------------------------"
+echo "----------------------"
+echo "Basic m2h tool testing"
+echo "----------------------"
 
-yarn md m2h markdown/tool/m2h --locale en-US
+yarn m2h markdown/tool/m2h --locale en-US
 diff -s testing/content/files/en-us/markdown/tool/m2h/index.html testing/content/files/en-us/markdown/tool/m2h/expected.html
-yarn md h2m markdown/tool/h2m --locale en-US
-diff -s testing/content/files/en-us/markdown/tool/h2m/index.md testing/content/files/en-us/markdown/tool/h2m/expected.md
