--- testing/content/files/en-us/markdown/tool/h2m/expected.md
@@ -1,54 +1,0 @@
----
-title: This is a test page
-slug: Markdown/Tool/H2M
-tags:
-  - Tag2
----
-`Markdown` is a nicer format than `HTML`.
-
-# Heading 1
-
-Nothing to say, really.
-
-## Heading 2
-
-Basic formatting such as **bold** and _emphasis_.
-
-Here's a section that simply mentions links like <https://www.peterbe.com> for example.
-
-### Heading 3
-
-A bullet point list:
-
-- One
-- Two
-- Three
-
-This is a paragraph with a link to [this page](/en-US/docs/Markdown/Tool/H2M).
-
-## Other stuff
-
-This is a paragraph with a link to [another page](/en-US/docs/Web).
-
-### A Definition List
-
-- Key
-  - : Value
-
-### Some code blocks
-
-```html
-  <canvas id="game" width="128" height="128"></canvas>
-```
-
-```js
-function foo() {
-  return undefined + 1;
-}
-```
-
-```css
-div#foo {
-  color: pink;
-}
-```
