--- testing/content/files/en-us/markdown/tool/h2m/index.html
@@ -1,41 +1,0 @@
----
-title: This is a test page
-slug: Markdown/Tool/H2M
-tags:
-  - Tag2
----
-<p><code>Markdown</code> is a nicer format than <code>HTML</code>.</p>
-<h1>Heading 1</h1>
-<p>Nothing to say, really.</p>
-<h2>Heading 2</h2>
-<p>Basic formatting such as <strong>bold</strong> and <em>emphasis</em>.</p>
-<p>Here's a section that simply mentions links like <a href="https://www.peterbe.com">https://www.peterbe.com</a> for example.</p>
-<h3>Heading 3</h3>
-<p>A bullet point list:</p>
-<ul>
-  <li>One</li>
-  <li>Two</li>
-  <li>Three</li>
-</ul>
-<p>This is a paragraph with a link to <a href="/en-US/docs/Markdown/Tool/H2M">this page</a>.</p>
-<h2>Other stuff</h2>
-<p>This is a paragraph with a link to <a href="/en-US/docs/Web">another page</a>.</p>
-<h3>A Definition List</h3>
-<dl>
-  <dt>Key</dt>
-  <dd>Value</dd>
-</dl>
-<h3>Some code blocks</h3>
-<pre class="brush: html">
-  &lt;canvas id="game" width="128" height="128"&gt;&lt;/canvas&gt;
-</pre>
-<pre class="brush: js">
-function foo() {
-  return undefined + 1;
-}
-</pre>
-<pre class="brush: css">
-div#foo {
-  color: pink;
-}
-</pre>
